requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
fastapi==0.123.4
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.7
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.13.0
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
posthog==5.4.0
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pypdf==6.4.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0
